fix(gui): re-review follow-ups — rejection latch, cursor edge
The Assigned-to latch now follows the live check out (a
fixed conflict can no longer resurrect the row when the same
combo is taken again later); the drag handle's onDisappear
restores whichever cursor the remaining hover state earns
instead of forcing arrow; the burst-candidate doc states the
real invariant (a chord-attempt overlap keeps the stash);
the recorder suite splits its release-burst tests into
ChordRecorderBurstTests.swift before the ceiling bites.
